lesson 12 (WIP): multiple ways to use entity command buffer (ECB) in parallel

+++ DotsMan/Assets/Scripts/Components/Collectable.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 [GenerateAuthoringComponent]
 public struct Collectable : IComponentData
 {
-	public float points;
+	public int points;
 
 }
